+++ app/Http/Controllers/Api/SizeController.php
@@ -20,16 +20,6 @@ class SizeController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
@@ -37,7 +27,15 @@ class SizeController extends Controller
      */
     public function store(Request $request)
     {
-
+        $this->validate(
+            $request,
+            [
+                'name' => 'required'
+            ],
+            [
+                'name.required' => 'Size name is empty!'
+            ]
+        );
         try{
             $data = [
                 "size_name" => $request->input("name"),
@@ -46,12 +44,12 @@ class SizeController extends Controller
             return response()->json([
                 'message' => "Create successfully!!",
                 'error' => false
-            ]);
+            ],200);
         }catch(Exception $e){
             return response()->json([
                 'message' => "Create failed. Try again!",
                 'error' => true
-            ]);
+            ],500);
         }
     }
 
@@ -75,16 +73,6 @@ class SizeController extends Controller
         }
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Size  $size
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Size $size)
-    {
-        //
-    }
 
     /**
      * Update the specified resource in storage.
@@ -93,23 +81,38 @@ class SizeController extends Controller
      * @param  \App\Models\Size  $size
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Size $size)
+    public function update(Request $request, $size_id)
     {
-
+        $this->validate(
+            $request,
+            [
+                'name' => 'required'
+            ],
+            [
+                'name.required' => 'Size name is empty!'
+            ]
+        );
         try {
             $data = [
                 'size_name' => $request->input('name'),
             ];
+            $size = Size::find($size_id);
+            if (is_null($size)) {
+                return response()->json([
+                    'message' => "Size is not exist",
+                    'error' => true
+                ],400);
+            }
             $size->update($data);
             return response()->json([
                 'message' => "Update size recored succesfully",
                 'error' => false
-            ]);
+            ],200);
         } catch (ModelNotFoundException $e) {
             return response()->json([
                 'message' => "Update size recored failed",
                 'error' => true
-            ]);
+            ],500);
         }
     }
 
@@ -119,19 +122,26 @@ class SizeController extends Controller
      * @param  \App\Models\Size  $size
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Size $size)
+    public function destroy($size_id)
     {
         try {
+            $size = Size::find($size_id);
+            if (is_null($size)) {
+                return response()->json([
+                    'message' => "Size is not exist",
+                    'error' => true
+                ],400);
+            }
             $size->delete();
             return response()->json([
                 'message' => "Delete successfully!",
                 'error' => false
-            ]);
+            ],200);
         } catch (Exception $e) {
             return response()->json([
                 'message' => "Delete failed! Try again",
                 'error' => true
-            ]);
+            ],500);
         }
     }
 }
